agregado detAll categoria

+++ src/main/java/ecommerce/eco/model/entity/Review.java
@@ -20,7 +20,7 @@ public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int score; // estrellas o puntaje
+    private double score; // estrellas o puntaje
     private String username; // nombre del usuario
     private String comment; // comentario
     @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
